Revert "Merge pull request #6 from primitivefinance/hotfix/clean"
This reverts commit 7e9e00418e7ccb1da05482f268175836af98474d, reversing
changes made to 78a8e64b7618e9199203ab84042876c580ae1e90.

+++ packages/primitive-contracts/contracts/Primitives.sol
@@ -2,12 +2,19 @@
 
 pragma solidity ^0.6.2;
 
+/**
+ * @title Library of Primitives
+ * @author Primitive
+ */
 library Primitives {
+    /*
+     * @notice Vanilla Option.
+     */
     struct Option {
         address underlyingToken;
         address strikeToken;
-        uint base;
-        uint quote;
-        uint expiry;
+        uint256 base;
+        uint256 quote;
+        uint256 expiry;
     }
 }
