+++ packages/primitive-contracts/contracts/option/interfaces/IFlash.sol
@@ -1,11 +1,21 @@
 // SPDX-License-Identifier: MIT
 
+
+
+
+
+
+
+
+
+
+
 pragma solidity ^0.6.2;
 
 interface IFlash {
     function primitiveFlash(
         address receiver,
-        uint outUnderlyings,
+        uint256 outUnderlyings,
         bytes calldata data
     ) external;
 }
