+++ packages/primitive-contracts/contracts/option/interfaces/IRedeem.sol
@@ -1,5 +1,15 @@
 // SPDX-License-Identifier: MIT
 
+
+
+
+
+
+
+
+
+
+
 pragma solidity ^0.6.2;
 
 interface IRedeem {
@@ -9,7 +19,7 @@ interface IRedeem {
 
     function factory() external view returns (address);
 
-    function mint(address user, uint amount) external;
+    function mint(address user, uint256 amount) external;
 
-    function burn(address user, uint amount) external;
+    function burn(address user, uint256 amount) external;
 }
