+++ packages/primitive-contracts/contracts/option/primitives/Redeem.sol
@@ -12,7 +12,7 @@ import { ERC20, IERC20 } from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import { SafeMath } from "@openzeppelin/contracts/math/SafeMath.sol";
 
 contract Redeem is IRedeem, ERC20 {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     address public override factory;
     address public override optionToken;
@@ -32,12 +32,12 @@ contract Redeem is IRedeem, ERC20 {
         redeemableToken = _redeemableToken;
     }
 
-    function mint(address to, uint amount) external override {
+    function mint(address to, uint256 amount) external override {
         require(msg.sender == optionToken, "ERR_NOT_VALID");
         _mint(to, amount);
     }
 
-    function burn(address to, uint amount) external override {
+    function burn(address to, uint256 amount) external override {
         require(msg.sender == optionToken, "ERR_NOT_VALID");
         _burn(to, amount);
     }
