+++ packages/primitive-contracts/contracts/test/CTokenLike.sol
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: MIT
 
+
+
 pragma solidity ^0.6.2;
 
 import "./ICToken.sol";
@@ -7,7 +9,7 @@ import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract CTokenLike is ICToken, ERC20 {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     address public override underlying;
 
@@ -19,14 +21,21 @@ contract CTokenLike is ICToken, ERC20 {
         underlying = _underlying;
     }
 
-    function mint(uint mintAmount) external override returns (uint) {
-        require(IERC20(underlying).balanceOf(msg.sender) >= mintAmount, "ERR_BAL_UNDERLYING");
+    function mint(uint256 mintAmount) external override returns (uint256) {
+        require(
+            IERC20(underlying).balanceOf(msg.sender) >= mintAmount,
+            "ERR_BAL_UNDERLYING"
+        );
         _mint(msg.sender, mintAmount);
         IERC20(underlying).transferFrom(msg.sender, address(this), mintAmount);
-        return uint(0);
+        return uint256(0);
     }
 
-    function redeemUnderlying(uint redeemAmount) external override returns (uint) {
+    function redeemUnderlying(uint256 redeemAmount)
+        external
+        override
+        returns (uint256)
+    {
         require(balanceOf(msg.sender) >= redeemAmount, "ERR_BAL_CTOKEN");
         _burn(msg.sender, redeemAmount);
         bool success = IERC20(underlying).transfer(msg.sender, redeemAmount);
@@ -35,7 +44,12 @@ contract CTokenLike is ICToken, ERC20 {
         } else return 1;
     }
 
-    function balanceOfUnderlying(address owner) external override view returns (uint) {
+    function balanceOfUnderlying(address owner)
+        external
+        override
+        view
+        returns (uint256)
+    {
         return balanceOf(owner);
     }
 }
